Extract LLM chat endpoints into a standalone Flask blueprint
Move all AI chat functionality out of the main application file and into
chat_blueprint.py, following the same plugin pattern used by project_views
and dataset_views.

- chat_blueprint.py: contains CHAT_MODEL, CHAT_TOOL_DEFS, build_system_prompt,
  execute_chat_tool, and the /chat + /api/chat routes via create_blueprint()
- execute_chat_tool receives get_entity_graph_nx via the shared helpers dict
- Routes use current_app.crucible_client for Crucible API calls
- crucible_graph_explore_flask_app.py: removed ~300 lines of chat code; the
  blueprint is registered via app.register_blueprint(chat_blueprint.create_blueprint(...))
